Fixed flakey directive definition.
Whether our directive to force focus onto the gamefield got defined
properly or not depended on the browser's load order for
scripts. Sometimes we tried to define the directive before angular
finished loading.

Moved the directive definition over to yasw.js, where things are
guaranteed to happen in the correct order.